pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "ai-rag-task"
version = "0.1.0"
description = "RAG system for resume analysis"

dependencies = [
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-text-splitters>=0.0.1",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.0.20",
    "chromadb>=0.4.0",
    "openai>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pypdf>=3.15.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "scikit-learn>=1.3.0",
    "rank-bm25>=0.2.2",
]

[project.scripts]
resume-rag = "resume_rag.main:main"